Pull TimeVerifier inside ScheduleTimeField
It's rather tightly coupled already and I want to reuse some functions
between the two classes for removing the leading underscore in the time.
